Moved types starting with 'Generic' into generics.rs.
